\name{cltbook}
\alias{cltbook}
\docType{data}
\title{Central Limit Theorem Workbook: html source}
\description{
This is an internal R-represenation of a workbook which can be written
to the users current working directory
}
\usage{data(cltbook)}
\format{
When written, this file is in .Rpad format, i.e. a standard html file
with R script sections
}

\source{
Based on materials used by Julian Stander
}
\references{
Perhaps some references need to be added
}
\examples{
data(cltbook)
## writeLines(cltbook, "clt.Rpad")
}
\keyword{datasets}
