\name{Rbooks-package}
\alias{Rbooks-package}
\alias{Rbooks}
\docType{package}
\title{
R-books: interactive learning materials using R in the background
}
\description{
There are tens/hundreds of thousands of rather cute electronic learning tools occupying a very modest fraction of cyberspace.   The aim of R-books is to figure out how best to use R to produce interactive course-notes / interactive learning materials.   The emphasis is on the notes rather than the R - currently the Rpad library is used to provide the interface between the user and R but that may change
}
\details{
\tabular{ll}{
Package: \tab Rbooks\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2008-04-03\cr
License: \tab GNU public licence version 2 or later\cr
}
}
\author{
Paul Hewson

Maintainer: Who to complain to <paul.hewson@plymouth.ac.uk>}
\references{
There is a lot of literature to add.   The work has been funded by an
MSOR grant.
At the moment there is (a) something horribly wrong with the image rendering and (b) getting the buttons to execute code within a node rather than on the whole page is taking a bit of figuring out.
}

\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-Rpad]{<pkg>}}
}
\examples{
##clt()
}
