\name{replacenotes}
\alias{replacenotes}
\title{replacenotes}
\description{
This is a slightly scary function which writes files into your user workspace.
}
\usage{
replacenotes()
}
\details{
  By default, the function will use a version of the course notes it finds in the working directory.   Setting overwrite = TRUE will copy a version from the package, along with all the images
}
\value{
  Nothing is returned from this function.   The side effect is that if you happened to have a load of files cltimg1.jpeg .. cltimg24.jpeg in your R working directory they will get over-written.
}
\references{There will be a lot of references added soon}
\author{Paul Hewson}
\note{This work has been funded by an MSOR grant}

\examples{
#library(Rbooks)
#data(cltbook)
#clt()


}

\keyword{misc}

